Add selection effects file.
This file contains all of the computations necessary to get the selection effects for the pulsar. To get the selection_effects (Smin, flux, etc.) this file will be the one that needs to be run. It takes the survey name, luminosity, and pulsar data file as the inputs when run and will create a csv output file with all of the computed values.