Start to add eccentricity considerations
I started writting a function to take care of considerations for eccentricity. I am not sure if we will need all of the equations in it and a few of them are not yet properly written, more edits will be necessary to make it a working function. It may not be fesable to use it due to the integrals in a few of the equations, I just wanted to get the equations down so that it could be modified until it is usable.