Check units and update I/P and O/P file format
I checked the units of most of the selection_effects functions and made corrections as necessary. A few of the units still don't seem quite right, specifically DM_0 and the eccentricity equations still need ironed out.

I also updated the input and output file format and modified a few functions to comply with this. Much of this was modifying how the binpulBothFile reads in the inputs and prints out the outputs as well as how the S_min and f_beaming functions where called.

Finally, I included some extra comments in the README file about units.