Modify binpulsBothFile for command line args
I modified the binpulsBothFile so that it can take command line arguments rather than asking for inputs from the user. This way it can be submitted to the batch system or on a local machine without too much trouble. In doing this, I needed to modify the survey names in survey.py because there where two named SKA. They have been renamed to SKA_2100 and SKA_300 depending on the integration time listed, all other values of the two surveys are the same.

I continued working trying to get the eccentricity function working for future use. I am not sure that it can be used for this application due to the nature of some of the variables in the equations and how they are defined by the paper.