Mark spots for review.
All of the pieces of the code that need review are in selection_effects.py. They should be marked with a comment starting with DEBATRI.

I marked the places for review within the eccentricity selection effects function with NOTE and added a list of issues in the documentation (marked with DEBATRI) instead because I am not sure that the function will be usable.

I also commented out the lines in the binpulsBothFile that involved the eccentricity function so that it is still usable without having the function working.