Change binpulBothFile to get T_sky_test
I edited the way that the ouptut files are created and written to hopefully be more efficient and clearer. Instead of adding to the input, I copied the input dataframe and appended columns to it for the output dataframe. Also, I created a T_sky_test dataframe that will output a csv file called T_sky_test containing the orignal inputs along with T_sky, S_min1*Area, S_min2*Area, Area.

A few edits were made to selection effects to modify the output and remove any uncessesary functions called.