Delete psrEvolve_input_dns.dat
I am removing this file from the repository because it does not appear that it will be needed to test/ create the output. We have the Debatri_Jorie_input_dnsCreated.dat which contains the same type of data in a different distribution. This file also has output counterparts which we can compare the python output to.