Make corrections to the files
I made some corrections and additions to the binpulsBothFile and selection_effects files in the hope of getting an output similar to what we get from the C code. As far as I can tell, the main issue is something within the We_fnct becoming too large for the returned value to be comparable to the period of the pulsar.