Solve S_min nan issue!
I fixed the issue with the S_min values returning nan. The issue was that pygedm (to get DM) expected the galactic coordinates to be in degree but they were given in radians. However, the returned values still do not match up with the expected output so more examination is necessary.