Update comment about units.
I added more comments about units to double check that we are using the correct ones in the equations so that there are few mix-ups.

I also added a extra input to the gal_cart cart2gal function which allows you to convert the cartisian x, y, z to galactic coordinates (l, b, d) where l and b are in degrees rather than radians.