Include multiple W_i functions
I updated the S_min_funct in selection_effects.py to calculate S_min and SNR for each of the three W_i functions that I have found. We will use this to determine which function for W_i we want to use in the final result.

I also updated the binpulsBothFile so that it will include each S_min and SNR in the final output file for the pulsar data. I updated the T_sky output so it will also include the multiple S_min values.