Modify T_sky function.
After a little bit of looking around, I found a paper on the healpix projection that the skytemp file uses. It looks like they use colatitude instead of latitude in their coordinates. I modified the function to account for this. I will test this and make modifications if needed. This may not work because skytempy may already be accounting for this.