Start to clean up files.
I updated the README.md file and added information about how to run the code as well as what the outputs are.

I removed some comments that are no longer necessary from the selection_effects.py, binpulsBothFile.py, survey.py, and gal_cart.py files.

I cleaned up and removed some of the uncessary print statements that had been commented out in the binpulsBothFile.py and selection_effects.py files.

I moved the eccentricity_fnct and E_fnct to a seperate file not included in the public repository. It is not currently working so it does not need to be included in the selection_effects.py file until it is able to run.