Update info in the README.md file and add new functs
I added a new function to the selection_effects.py file which tests for the radio visibility cutoff of DNS and NSBH systems. It returns two alt_det values, one for each object in the system.

I added an optional line_cutoff argument which takes a boolean variable. If True, the pulsar's whose pdot value falls below the deathline will have it's period removed from the dataframe, preventing it from being calculated. This needs to be tested!!

I added the new argument and it's info to the README.md file.