Test COSMIC input
I wrote a new python file which converts all the values given in the COSMIC output to the units and file type that we want when running them through the selection effects code. I have also included the converted COSMIC output so it can be run throug the selection effects code.